Refactoring login
